+++ CHANGELOG.md
@@ -14,6 +14,7 @@ All notable changes to this project will be documented in this file.
   - Force IPv4-only connections for gNMI tunnel client and fix TLS credential handling
 - Client
   - Support simultaneous unicast and multicast tunnels in doublezerod
+  - Support probing interfaces marked as tunnel endpoints in doublezerod
 - SDK
   - Go SDK can now perform batch writes to device.health and link.health as per rfc12